[CWS] split raw packet unmarshaller in build gated file
